backend/pyproject.toml
[project]
name = "trustlayer-backend"
version = "0.1.0"
description = "TrustLayer — C2PA + behavioral attestation + soft binding"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "cryptography>=43.0.0",
    "pillow>=10.4.0",
    "numpy>=2.0.0",
    "scipy>=1.14.0",
    "python-multipart>=0.0.12",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
